backend/requirements.txt
uvicorn>=0.26.0
fastapi>=0.109.0
pydantic_settings>=2.1.0
requests>=2.23.0
etlrules>=0.3.2
